# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/promo_bundles:
    get:
      tags:
        - PromoBundles
      summary: List PromoBundles
      description: Get a paginated list of PromoBundles with optional filtering and sorting
      operationId: listPromoBundles
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoBundleList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PromoBundles
      summary: Create PromoBundle
      description: Create a new PromoBundle
      operationId: createPromoBundle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromoBundleRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoBundle'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/promo_bundle_gifts:
    get:
      tags:
        - PromoBundleGifts
      summary: List PromoBundleGifts
      description: Get a paginated list of PromoBundleGifts with optional filtering and sorting
      operationId: listPromoBundleGifts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoBundleGiftList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PromoBundleGifts
      summary: Create PromoBundleGift
      description: Create a new PromoBundleGift
      operationId: createPromoBundleGift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromoBundleGiftRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoBundleGift'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PromoBundle:
      type: object
      required:
        - id
        - company_id
        - title
        - priority
        - match_type
        - reward
        - currency
        - min_order_amount
        - stackable
        - valid_from
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 140
        priority:
          type: integer
          format: int32
        match_type:
          type: string
        required_distinct:
          nullable: true
          type: integer
          format: int32
        reward:
          type: string
        discount_percentage:
          nullable: true
          type: string
        discount_amount:
          nullable: true
          type: string
        reward_item_id:
          nullable: true
          type: string
          format: uuid
        reward_qty:
          nullable: true
          type: string
        currency:
          type: string
        min_order_amount:
          type: string
        stackable:
          type: boolean
        valid_from:
          type: string
          format: date-time
        valid_to:
          nullable: true
          type: string
          format: date-time
        status:
          type: string
        metadata:
          type: object
...
    CreatePromoBundleRequest:
      type: object
      required:
        - company_id
        - title
        - priority
        - match_type
        - reward
        - currency
        - min_order_amount
        - stackable
        - valid_from
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 140
        priority:
          type: integer
          format: int32
        match_type:
          type: string
        required_distinct:
          nullable: true
          type: integer
          format: int32
        reward:
          type: string
        discount_percentage:
          nullable: true
          type: string
        discount_amount:
          nullable: true
          type: string
        reward_item_id:
          nullable: true
          type: string
          format: uuid
        reward_qty:
          nullable: true
          type: string
        currency:
          type: string
        min_order_amount:
          type: string
        stackable:
          type: boolean
        valid_from:
          type: string
          format: date-time
        valid_to:
          nullable: true
          type: string
          format: date-time
        status:
          type: string
        metadata:
          type: object
...
    PromoBundleList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PromoBundle'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PromoBundleGift:
      type: object
      required:
        - id
        - company_id
        - bundle_id
        - gift_item_id
        - gift_qty
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        bundle_id:
          type: string
          format: uuid
        gift_item_id:
          type: string
          format: uuid
        gift_qty:
          type: string
        metadata:
          type: object
...
    CreatePromoBundleGiftRequest:
      type: object
      required:
        - company_id
        - bundle_id
        - gift_item_id
        - gift_qty
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        bundle_id:
          type: string
          format: uuid
        gift_item_id:
          type: string
          format: uuid
        gift_qty:
          type: string
        metadata:
          type: object
...
    PromoBundleGiftList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PromoBundleGift'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []